Fix linux build

+++ src/utils/widgets/flowlayout.cpp
@@ -47,7 +47,7 @@ int FlowLayout::verticalSpacing() const
 
 int FlowLayout::count() const
 {
-	return itemList.size();
+	return (int)itemList.size();
 }
 
 QLayoutItem *FlowLayout::itemAt(int index) const
